+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/common/repository/CertificateRepoContainer.kt
@@ -1,10 +1,10 @@
 package de.rki.coronawarnapp.covidcertificate.common.repository
 
-import de.rki.coronawarnapp.covidcertificate.common.recycle.Recyclable
+import de.rki.coronawarnapp.reyclebin.common.Recyclable
 import de.rki.coronawarnapp.util.qrcode.coil.CoilQrCode
 import java.util.Locale
 
-interface CertificateRepoContainer : Recyclable {
+interface CertificateRepoContainer: Recyclable {
     val containerId: CertificateContainerId
 
     /**
