+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/test/core/storage/types/BaseTestCertificateData.kt
@@ -1,13 +1,13 @@
 package de.rki.coronawarnapp.covidcertificate.test.core.storage.types
 
 import de.rki.coronawarnapp.covidcertificate.common.certificate.CwaCovidCertificate
-import de.rki.coronawarnapp.covidcertificate.common.recycle.Recyclable
+import de.rki.coronawarnapp.reyclebin.common.Recyclable
 import org.joda.time.Instant
 
 /**
  * Common data for test certificates, independent of whether they were retrieved or scanned.
  */
-sealed class BaseTestCertificateData : Recyclable {
+sealed class BaseTestCertificateData: Recyclable {
     abstract val identifier: String
     abstract val registeredAt: Instant
     abstract val certificateReceivedAt: Instant?
